start to cleaning the house
