+++ src/generator/mod.rs
@@ -1,1 +1,2 @@
 pub mod generator;  
+pub mod generate_tx;
