+++ src/main.rs
@@ -1,7 +1,6 @@
 use clap::{Parser, Subcommand};
 mod generator;
 use generator::generator::Generator;
-
 #[derive(Parser)]
 #[command(version, about, long_about = None)]
 struct Cli {
@@ -12,8 +11,12 @@ struct Cli {
 #[derive(Subcommand)]
 enum Commands {
     Numberoftxs {
-        input1: i32 // Self anotation, for now the only required input is the number of txs thats define if you take a tx or a block
+        txscount: i32 
+                    // Self anotation, for now the only required input is the number of txs thats define if you take a tx or a block
                     // but in the future the user will need input the campus to break and the kind of transaction... 
+    },
+    CampusToBreak{
+        listedcampus:Vec<String>
     }
 
 }
@@ -21,9 +24,13 @@ enum Commands {
 fn main() {
     let cli = Cli::parse();
     match cli.command {
-        Commands::Numberoftxs { input1 } => {
-            let result = Generator::generate_from_input(input1);
-            println!("{}", result);
+        Commands::Numberoftxs { txscount } => {
+            let input_txs_count = Generator::generate_from_input(txscount);
+            println!("{}", input_txs_count);
+        },
+        Commands::CampusToBreak { listedcampus } =>{
+            let input_listedcampus = Generator::proces_flags_to_broke(listedcampus);
+            println!("{}", input_listedcampus)
         }
     }
 }
